requirements.txt
supervision
openai
opencv-python
numpy
requests
gradio==4.44.1

vertexai
pillow
google.generativeai

pynput
pyautogui